refactor: 로또 생성 및 구입 로직을 별도의 LottoMachine 클래스로 분리
• 로또를 생성하고 판매하는 역할을 가진 LottoMachine 클래스를 도입하였습니다.
• 관련 로직들을 LottoMachine 클래스로 분리하였습니다.
• 하드코딩 값을 상수 처리하였습니다.
• 수량만큼 로또를 생성하는 로직을 스트림으로 개선하였습니다.
