+++ src/main/java/lotto/domain/Lotto.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import java.util.Arrays;
 import java.util.List;
 
 public class Lotto {
@@ -8,7 +7,6 @@ public class Lotto {
     private static final int LOTTO_NUMBERS_SIZE = 6;
     private static final int MIN_LOTTO_NUMBER = 1;
     private static final int MAX_LOTTO_NUMBER = 45;
-    private static final String LOTTO_NUMBER_DELIMITER = ",";
     private static final String SIZE_EXCEPTION = "로또는 %d개의 번호로 이뤄져야 합니다.";
     private static final String DUPLICATE_EXCEPTION = "로또에 같은 번호가 존재합니다.";
     private static final String RANGE_EXCEPTION = "로또 번호는 %d부터 %d 사이의 숫자여야 합니다.";
@@ -18,14 +16,6 @@ public class Lotto {
     public Lotto(List<Integer> numbers) {
         validate(numbers);
         this.numbers = sort(numbers);
-    }
-
-    public static Lotto makeWinningLotto(String winningNumbers) {
-        List<Integer> numbers = Arrays.stream(winningNumbers.split(LOTTO_NUMBER_DELIMITER))
-                .map(String::trim)
-                .map(Integer::parseInt)
-                .toList();
-        return new Lotto(numbers);
     }
 
     private void validate(List<Integer> numbers) {
